Implement per-guardian health and targeting for BossGuardian
Adds individual health pools and hitboxes for each BossGuardian, allowing player attacks and projectiles to target and damage each guardian separately. Introduces BossGuardianEnemyAdapter to expose guardians as Enemy instances, updates Enemy interface to use float health/damage, and adjusts related skills and tentacle logic. Also includes new boss skill assets, level and save data updates, and minor UI and debug improvements.

+++ Chromashift/core/src/main/java/com/jjmc/chromashift/environment/enemy/Enemy.java
@@ -11,13 +11,13 @@ public interface Enemy {
      * Apply damage to this enemy.
      * @param damage The amount of damage to apply
      */
-    void takeDamage(int damage);
+    void takeDamage(float damage);
     
     /**
      * Get the current health of this enemy.
      * @return Current health value
      */
-    int getHealth();
+    float getHealth();
     
     /**
      * Check if this enemy is still alive.
